requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
Pillow==11.1.0
